requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# Pydantic
pydantic==2.10.5
pydantic-settings==2.7.1
typing-extensions==4.12.2

# LangChain Stack - Compatible versions for Python 3.12
langchain==0.3.15
langchain-core==0.3.31
langchain-openai==0.2.14
langchain-community==0.3.15
langgraph==0.2.62

# OpenAI
openai==1.59.8

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Text Processing
tiktoken==0.8.0
